Update README.md
It will ask

| Would you like to continue with one of these lessons?

The list of possible lessons to continue with, depends on what you've been working on before. So, for me, this list is different from this list. I suggest to write this more generic.

Also, I thought we were running the IMBRSea-R-Working-Group-SummeR lessons, not the ImbrseaR course (this is the old version, right?) ...